Add general section and fix header buttons
I added the general section to the info page. So far only the name input
works. I don't really like passing functions around just to change an
input value. I also realized that you could keep pressing one of the
header buttons and it would keep changing pages. I added an IF statement
to check if the IDs are different. If the ID is the same as the current
state, then it won't switch pages. Also, the buttons IDs are from the
Page enum.