Finish skills section and add certifcation section
I finished adding style to the skills section.

I also modified the way it moves the skills up and down. Since I was
going to make the sections moveable as well, I made a function that
handles moving from one index to another index in an array. It will
return a new array that has the elements in the correct order. This is
the moveTo function in the array-functions.js file.

I added the same arrow buttons to the sections below general. I want the
general section to always be at the top. Those buttons use the same
function (moveTo) to modify the order array.

The order array is an array that is used to map the sections in the info
and preview screen. It stores a value from the Section enum. A function
inside of Info.js called getSection uses a switch statement to return
the applicable component based on the stored value.

A similar function to getSection will need to be made in the Preview.js
component to ensure the proper order is displayed on the resume.